% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchaeoData-package.R
\docType{package}
\name{ArchaeoData-package}
\alias{ArchaeoData-package}
\alias{ArchaeoData}
\title{ArchaeoData: Datasets for 'ArchaeoPhases'}
\description{
Datasets for chronological modelling with 'ArchaeoPhases'. This package provides models and data to reproduce results from 'ArchaeoPhases' examples and vignettes.
}
\details{
\tabular{ll}{
\strong{Version:} \tab 0.1.0 \cr
\strong{License:} \tab GPL-3 \cr
}

Laboratoire de Mathématiques Jean Leray (UMR 6629)\cr
2, rue de la Houssinière\cr
BP 92208\cr
F-44322 Nantes Cedex 3\cr
France
}
\seealso{
Useful links:
\itemize{
  \item \url{https://archaeostat.github.io/archaeodata/}
  \item Report bugs at \url{https://github.com/ArchaeoStat/ArchaeoData/issues}
}

}
\author{
\strong{Maintainer}: Anne Philippe \email{anne.philippe@univ-nantes.fr} (\href{https://orcid.org/0000-0002-5331-5087}{ORCID})

Authors:
\itemize{
  \item Nicolas Frerebeau \email{nicolas.frerebeau@u-bordeaux-montaigne.fr} (\href{https://orcid.org/0000-0001-5759-4944}{ORCID})
  \item Thomas S. Dye (\href{https://orcid.org/0000-0001-8116-782X}{ORCID})
}

}
\keyword{internal}
