% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArchaeoPhases.dataset-package.R
\docType{package}
\name{ArchaeoPhases.dataset-package}
\alias{ArchaeoPhases.dataset-package}
\alias{_PACKAGE}
\alias{ArchaeoPhases.dataset}
\title{ArchaeoPhases.dataset: Datasets for 'ArchaeoPhases'}
\description{
Datasets for chronological modelling with 'ArchaeoPhases'. This package provides models and data to reproduce results from 'ArchaeoPhases' examples and vignettes.
}
\details{
\tabular{ll}{
\strong{Package:} \tab ArchaeoPhases.dataset \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.0.0.9000 \cr
\strong{License:} \tab GPL-3 \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ArchaeoStat/ArchaeoPhases.dataset}
  \item Report bugs at \url{https://github.com/ArchaeoStat/ArchaeoPhases.dataset/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Thomas S. Dye \tab \emph{University of Hawai'i at Manoa, USA} \cr
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Anne Philippe \tab \emph{Université de Nantes, France} \cr
}

\strong{Package maintainer}

Anne Philippe\cr
\email{anne.philippe@univ-nantes.fr}

Laboratoire de Mathématiques Jean Leray (UMR 6629)\cr
2, rue de la Houssinière\cr
BP 92208\cr
F-44322 Nantes Cedex 3\cr
France
}
\keyword{internal}
