% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{burials}
\alias{burials}
\title{Anglo-Saxon Female Burials with Beads}
\format{
A \code{\link{data.frame}} with 5,000 rows and 77 columns:
\describe{
\item{\code{UB-6041 (CasD182)}}{Date of burial CasD182.}
\item{\code{UB-6038 (CasD183)}}{Date of burial CasD183.}
\item{\code{UB-4960 (BuD391B)}}{Date of burial BuD391B.}
\item{\code{UB-4959 (BuD391A)}}{Date of burial BuD391A.}
\item{\code{UB-4511 (EH090)}}{Date of burial EH090.}
\item{\code{UB-4512 (EH091)}}{Date of burial EH091.}
\item{\code{MelSG077}}{Date of burial MelSG077.}
\item{\code{UB-4885 (MelSG078)}}{Date of burial MelSG078.}
\item{\code{UB-4884 (MelSG079)}}{Date of burial MelSG079.}
\item{\code{UB-4882 (MelSG080)}}{Date of burial MelSG080.}
\item{\code{UB-6476 (BuD339)}}{Date of burial BuD339.}
\item{\code{UB-4734 (MH105c)}}{Date of burial MH105c.}
\item{\code{UB-4732 (MH094)}}{Date of burial MH094.}
\item{\code{UB-4890 (MelSG075)}}{Date of burial MelSG075.}
\item{\code{UB-4728 (MH064)}}{Date of burial MH064.}
\item{\code{UB-4733 (MH095)}}{Date of burial MH095.}
\item{\code{UB-6473 (BuD250)}}{Date of burial BuD250.}
\item{\code{UB-4735 (Ber022)}}{Date of burial Ber022.}
\item{\code{UB-4739 (Ber134/1)}}{Date of burial Ber134/1.}
\item{\code{UB-4836 (WG27)}}{Date of burial WG27.}
\item{\code{UB-6472 (BuD222)}}{Date of burial BuD222.}
\item{\code{UB-6037 (CasD134)}}{Date of burial CasD134.}
\item{\code{UB-4888 (MelSG089)}}{Date of burial MelSG089.}
\item{\code{UB-6040 (CasD053)}}{Date of burial CasD053.}
\item{\code{UB-4707 (EH079)}}{Date of burial EH079.}
\item{\code{UB-6035 (CasD096)}}{Date of burial CasD096.}
\item{\code{UB-4975 (AstCli12)}}{Date of burial AstCli12.}
\item{\code{UB-4984 (Lec018)}}{Date of burial Lec018.}
\item{\code{UB-4835 (ApD134)}}{Date of burial ApD134.}
\item{\code{UB-4729 (MH068)}}{Date of burial MH068.}
\item{\code{UB-6034 (CasD120)}}{Date of burial CasD120.}
\item{\code{UB-4705 (WHes123)}}{Date of burial WHes123.}
\item{\code{UB-6033 (WHes113)}}{Date of burial WHes113.}
\item{\code{UB-4709 (EH014)}}{Date of burial EH014.}
\item{\code{UB-4708 (EH083)}}{Date of burial EH083.}
\item{\code{UB-5208 (ApD107)}}{Date of burial ApD107.}
\item{\code{UB-4077 (But4275)}}{Date of burial But4275.}
\item{\code{UB-4965 (ApD117)}}{Date of burial ApD117.}
\item{\code{UB-4889 (MelSG069)}}{Date of burial MesSG069.}
\item{\code{UB-4963 (SPTip208)}}{Date of burial SPTip208.}
\item{\code{UB-6032 (SPTip073)}}{Date of burial SPTip073.}
\item{\code{UB-6036 (CasD013)}}{Date of burial CasD013.}
\item{\code{UB-4887 (MelSG082)}}{Date of burial MeSG082.}
\item{\code{UB-4964 (Cod30)}}{Date of burial Cod30.}
\item{\code{UB-4883 (MelSG095)}}{Date of burial MelSG095.}
\item{\code{UB-4042 (But1674)}}{Date of burial But1674.}
\item{\code{UB-4552 (MaDE3)}}{Date of burial MaDE3.}
\item{\code{UB-4507 (Lec187)}}{Date of burial Lec187.}
\item{\code{UB-4706 (WHes118)}}{Date of burial WHes118.}
\item{\code{UB-4502 (Lec138)}}{Date of burial Lec138.}
\item{\code{UB-4504 (Lec179)}}{Date of burial Lec179.}
\item{\code{UB-4910 (BloodH22)}}{Date of burial BloodH22.}
\item{\code{UB-4506 (Lec172/2)}}{Date of burial Lec172/2.}
\item{\code{MaDE1 & E2}}{Date of burial MaDE1 & E2.}
\item{\code{UB-4554 (MaDF2)}}{Date of burial MaDF2.}
\item{\code{UB-4549 (MaDC7)}}{Date of burial MaDC7.}
\item{\code{UB-4553 (MaDD10)}}{Date of burial MaDD10.}
\item{\code{UB-6042 (CasD088)}}{Date of burial CasD088.}
\item{\code{UB-4501 (Lec014)}}{Date of burial Lec014.}
\item{\code{UB-4503 (Lec148)}}{Date of burial Lec148.}
\item{\verb{SUERC-51539 (ERL G353)}}{Date of burial ERL G353.}
\item{\verb{SUERC-51548 (ERL G210)}}{Date of burial ERL G210.}
\item{\verb{SUERC-51553 (ERL G116)}}{Date of burial ERL G116.}
\item{\verb{SUERC-39108 ERLK G322}}{Date of burial ERLK G322.}
\item{\verb{SUERC-39109 ERL G362}}{Date of burial ERL G362.}
\item{\verb{SUERC-39112 ERL G405}}{Date of burial ERL G405.}
\item{\verb{SUERC-51560 ERL G038}}{Date of burial ERL G038.}
\item{\verb{SUERC-39091 (ERL G003)}}{Date of burial ERL G003.}
\item{\verb{SUERC-39092 (ERL G005)}}{Date of burial ERL G005.}
\item{\verb{SUERC-39113 (ERL G417)}}{Date of burial ERL G417.}
\item{\verb{SUERC-51549 (ERL G195)}}{Date of burial ERL G195.}
\item{\verb{SUERC-51543 (ERL G281)}}{Date of burial ERL G281.}
\item{\verb{SUERC-51551 (ERL G193)}}{Date of burial ERL G193.}
\item{\verb{SUERC-51552 (ERL G107)}}{Date of burial ERL G107.}
\item{\verb{SUERC-39100 (ERL G266)}}{Date of burial ERL G266.}
\item{\verb{SUERC-51550 (ERL G254)}}{Date of burial ERL G254.}
\item{\verb{SUERC-39096 (ERL G112)}}{Date of burial ERL G112.}
}
}
\usage{
burials
}
\description{
Results of an OxCal calibration.
}
\seealso{
Other datasets: 
\code{\link{ChronoModelEvents}},
\code{\link{ChronoModelPhases}},
\code{\link{fishpond}},
\code{\link{ksarakil}}
}
\concept{datasets}
\keyword{datasets}
