% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChronoModelEvents}
\alias{ChronoModelEvents}
\title{Ksar'Akil Dates Calibrated by ChronoModel}
\format{
A \code{\link{data.frame}} with 30,000 rows and 17 variables:
\describe{
\item{\code{iter}}{Iteration of the MCMC algorithm}
\item{\code{V}}{Layer V.}
\item{\code{VI}}{Layer VI.}
\item{\code{XI}}{Layer XI.}
\item{\code{XII}}{Layer XII.}
\item{\code{XVI_1}}{Layer XVI 1.}
\item{\code{XVI_2}}{Layer XVI 2.}
\item{\code{XVI_3}}{Layer XVI 3.}
\item{\code{XVI_4}}{Layer XVI 4.}
\item{\code{XVII_1}}{Layer XVII 1.}
\item{\code{XVII_2}}{Layer XVII 2.}
\item{\code{XVII_3}}{Layer XVII 3.}
\item{\code{XVII_4}}{Layer XVII 4.}
\item{\code{XVIII}}{Layer XVIII.}
\item{\code{XIX}}{Layer XIX.}
\item{\code{XX}}{Layer XX.}
\item{\code{XXII}}{Layer XXII.}
}
}
\usage{
data(ChronoModelEvents)
}
\description{
Ksar'Akil Dates Calibrated by ChronoModel
}
\seealso{
Other datasets: 
\code{\link{ChronoModelPhases}},
\code{\link{burials}},
\code{\link{fishpond}},
\code{\link{ksarakil}}
}
\concept{datasets}
\keyword{datasets}
