% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChronoModelPhases}
\alias{ChronoModelPhases}
\title{Ksar'Akil Phases Calibrated by ChronoModel}
\format{
A \code{\link{data.frame}} with 30,000 rows and 9 variables:
\describe{
\item{\code{iter}}{Iteration of the MCMC algorithm.}
\item{\verb{EPI begin}}{Start date of EPI.}
\item{\verb{EPI end}}{End date of EPI.}
\item{\verb{UP begin}}{Start date of UP.}
\item{\verb{UP end}}{End date of UP.}
\item{\verb{Ahmarian begin}}{Start date of Ahmarian.}
\item{\verb{Ahmarian end}}{End date of Ahmarian.}
\item{\verb{IUP begin}}{Start date of IUP.}
\item{\verb{IUP end}}{End date of IUP.}
}
}
\usage{
data(ChronoModelPhases)
}
\description{
Ksar'Akil Phases Calibrated by ChronoModel
}
\seealso{
Other datasets: 
\code{\link{ChronoModelEvents}},
\code{\link{burials}},
\code{\link{fishpond}},
\code{\link{ksarakil}}
}
\concept{datasets}
\keyword{datasets}
