% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishpond}
\alias{fishpond}
\title{Calibration of a Fishpond Chronology}
\format{
A \code{\link{data.frame}} with 55,964 rows and 11 variables:
\describe{
\item{\code{iteration}}{Iteration of the MCMC algorithm.}
\item{\code{beta_2_Layer_II}}{End date of Layer II.}
\item{\code{theta_5_Layer_II}}{Age of dated event 5 in Layer II.}
\item{\code{theta_4_Layer_II}}{Age of dated event 4 in Layer II.}
\item{\code{theta_3_Layer_II}}{Age of dated event 3 in Layer II.}
\item{\code{theta_2_Layer_II}}{Age of dated event 2 in Layer II.}
\item{\code{alpha_2_Layer_II}}{Start date of Layer II.}
\item{\code{beta_1_Layer_III}}{End date of Layer III.}
\item{\code{theta_1_Layer_III}}{Age of dated event 1 in Layer III.}
\item{\code{alpha_1_Layer_III}}{Start date of Layer III.}
\item{\code{phi_1}}{Floating parameter.}
}
}
\usage{
fishpond
}
\description{
A data set containing information on the ages of two fishpond deposits.
}
\seealso{
Other datasets: 
\code{\link{ChronoModelEvents}},
\code{\link{ChronoModelPhases}},
\code{\link{burials}},
\code{\link{ksarakil}}
}
\concept{datasets}
\keyword{datasets}
