% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ksarakil}
\alias{ksarakil}
\title{Ksar'Akil Dates Calibrated by OxCal}
\format{
A \code{\link{data.frame}} with 1000 rows and 27 variables:
\describe{
\item{\code{Pass}}{iteration of the MCMC algorithm}
\item{\code{Ethelruda}}{Ethelruda.}
\item{\verb{start dated IUP}}{Start dated IUP.}
\item{\code{GrA-53000}}{GrA-5300.}
\item{\verb{end dated IUP}}{End dated IUP.}
\item{\verb{start Ahmarian}}{Start Ahmarian.}
\item{\code{GrA-57597}}{GrA-57597.}
\item{\code{GrA-53004}}{GrA-53004.}
\item{\code{GrA-57542}}{GrA-57542.}
\item{\code{GrA-54846}}{GrA-54846.}
\item{\code{GrA-57603}}{GrA-57603.}
\item{\code{GrA-57602}}{GrA-57602.}
\item{\code{GrA-53001}}{GrA-53001.}
\item{\code{Egbert}}{Egbert.}
\item{\code{GrA-54847}}{GrA-54847.}
\item{\code{GrA-57599}}{GrA-57599.}
\item{\code{GrA-57598}}{GrA-57598.}
\item{\code{GrA-57544}}{GrA-57544.}
\item{\verb{end Ahmarian}}{End Ahmarian.}
\item{\verb{start UP}}{Start UP.}
\item{\code{GrA-57545}}{GrA-57545.}
\item{\code{GrA-53006}}{GrA-53006.}
\item{\code{GrA-54848}}{GrA-54848.}
\item{\verb{end UP}}{End UP.}
\item{\verb{start EPI}}{Start EPI.}
\item{\code{GrA-53005}}{GrA-53005.}
\item{\verb{end EPI}}{End EPI.}
}
}
\usage{
data(ksarakil)
}
\description{
Ksar'Akil Dates Calibrated by OxCal
}
\seealso{
Other datasets: 
\code{\link{ChronoModelEvents}},
\code{\link{ChronoModelPhases}},
\code{\link{burials}},
\code{\link{fishpond}}
}
\concept{datasets}
\keyword{datasets}
